requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
brotlipy==0.7.0
celery==5.2.7
certifi # @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_884c889c-96af-444f-bd6d-daddb5e9a462ykj3l5n_/croots/recipe/certifi_1655968814730/work/certifi
cffi #@ file:///Users/ktietz/ci_310/cffi_1643911190082/work
charset-normalizer # @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
#cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5871d1ea-0250-4cd7-ac89-4b1e60514f5daqk8t0ow/croots/recipe/cryptography_1652101128666/work
Django==4.0.6
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-timezone-field==5.0
djangorestframework==3.13.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna # @ file:///tmp/build/80754af9/idna_1637925883363/work
kombu==5.2.4
pika==1.3.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
#pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
#pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
#PySocks @ file:///Users/ktietz/ci_310/pysocks_1643961536721/work
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests # @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b15e28ca-768e-4951-b4c6-8a7476e43183hrjpnkg9/croots/recipe/requests_1657734638595/work
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3 # @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e1a192c4-15be-4e72-95cc-219f6405cdb5nmvan_lb/croots/recipe/urllib3_1650637219436/work
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
